interface under design
I'm working on Alacast's interface.  I'll be its initial interface with GTK and more specifically Glade/libglade.  My 1st interface will just be to get a working usable version working as soon as possible.

But its just not as beautiful as I want Ala ast to be so for my release version I'll be creating my interface using Clutter or Pigment.  I prefer how Pigment's infaces look but I may end up using Clutter.  With Clutter's hooks into GStreamer & GTK I have a feeling I'll be more likely to get my over all look & feel more easilly using Clutter than I would with either Pigment or GTK alone.

I'll also be using my OPML podcast listings for my initial interface but Alacast's storage engine will be switched over to sqlite.  My plans are to implement my GTK interface using my OPML files.  Than once I impliment Alacast's 'import OPML' features & methods than I'll switch over to using sqlite as its primary storage engine.

I will have lirc support implement as soon as possible, meaning as soon as I actually get a remote setup & working on my system.

I feel very glad to finally be focusing on Alacast as my primary project *w00t.

Committed on: 2009-03-17 @ 09:10:58 AM

+++ data/xml/opml/Radiocasts/NPR/create_opmls.tcsh
@@ -1,6 +1,8 @@
 #!/bin/tcsh -f
-if ( -e 'index.opml' ) \
-	wget -O 'index.opml' 'http://www.npr.org/rss/podcast/podcast_directory.php?type=topic';
+set NPR_DIR="`dirname ${0}`/data/xml/opml/Radiocasts/NPR";
+cd "${NPR_DIR}"
+
+if ( -e 'index.opml' ) wget -O 'index.opml' 'http://www.npr.org/rss/podcast/podcast_directory.php?type=topic';
 
 # these regex's will make sure that each catagory is placed on one line for each catagory.
 ex '+1,$s/[\ \t]\+//g' '+1,$s/[\n\r]\+//g' '+s/.*<div class="topicForm">\(.*\)/\1/g' '+s/.*\(<span class="mainTopicLnk">\)/\r\1/g' '+1d' '+$s/<div class="spacer">.*$//g' '+wq' 'index.opml'
